@@ -12,52 +12,52 @@
   },
   "private": true,
   "dependencies": {
-    "@nrwl/nx-cloud": "^13.1.2",
-    "@reduxjs/toolkit": "^1.8.0",
+    "@nrwl/nx-cloud": "14.0.5",
+    "@reduxjs/toolkit": "1.8.1",
     "core-js": "^3.6.5",
-    "express": "4.17.1",
+    "express": "4.17.2",
     "pg": "^8.7.3",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^6.2.2",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
+    "react-is": "18.0.0",
+    "react-redux": "8.0.1",
+    "react-router-dom": "6.3.0",
     "regenerator-runtime": "0.13.7",
     "tslib": "^2.0.0"
   },
   "devDependencies": {
-    "@nrwl/cli": "13.4.6",
-    "@nrwl/cypress": "13.4.6",
-    "@nrwl/eslint-plugin-nx": "13.4.6",
-    "@nrwl/express": "13.4.6",
-    "@nrwl/jest": "13.4.6",
-    "@nrwl/linter": "13.4.6",
-    "@nrwl/node": "13.4.6",
-    "@nrwl/react": "13.4.6",
-    "@nrwl/tao": "13.4.6",
-    "@nrwl/web": "13.4.6",
-    "@nrwl/workspace": "13.4.6",
-    "@testing-library/react": "12.1.2",
-    "@testing-library/react-hooks": "7.0.2",
-    "@types/express": "4.17.0",
-    "@types/jest": "27.0.2",
+    "@nrwl/cli": "14.1.9",
+    "@nrwl/cypress": "14.1.9",
+    "@nrwl/eslint-plugin-nx": "14.1.9",
+    "@nrwl/express": "14.1.9",
+    "@nrwl/jest": "14.1.9",
+    "@nrwl/linter": "14.1.9",
+    "@nrwl/node": "14.1.9",
+    "@nrwl/react": "14.1.9",
+    "@nrwl/web": "14.1.9",
+    "@nrwl/workspace": "14.1.9",
+    "@testing-library/react": "13.1.1",
+    "@types/express": "4.17.13",
+    "@types/jest": "27.4.1",
     "@types/node": "16.11.7",
     "@types/pg": "^8.6.5",
-    "@types/react": "17.0.30",
-    "@types/react-dom": "17.0.9",
-    "@typescript-eslint/eslint-plugin": "~5.3.0",
-    "@typescript-eslint/parser": "~5.3.0",
+    "@types/react": "18.0.8",
+    "@types/react-dom": "18.0.3",
+    "@typescript-eslint/eslint-plugin": "5.18.0",
+    "@typescript-eslint/parser": "5.18.0",
     "cypress": "^9.1.0",
-    "eslint": "8.2.0",
+    "eslint": "8.12.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
-    "eslint-plugin-import": "2.25.2",
+    "eslint-plugin-import": "2.26.0",
     "eslint-plugin-jsx-a11y": "6.5.1",
-    "eslint-plugin-react": "7.27.0",
-    "eslint-plugin-react-hooks": "4.2.0",
-    "jest": "27.2.3",
-    "prettier": "^2.3.1",
-    "react-test-renderer": "17.0.2",
-    "ts-jest": "27.0.5",
+    "eslint-plugin-react": "7.29.4",
+    "eslint-plugin-react-hooks": "4.5.0",
+    "jest": "27.5.1",
+    "nx": "14.1.9",
+    "prettier": "2.5.1",
+    "react-test-renderer": "18.1.0",
+    "ts-jest": "27.1.4",
     "typescript": "~4.4.3"
   }
 }
